requirement.txt
asn1crypto==1.5.1
attrs==23.1.0
boto3==1.33.9
botocore==1.33.9
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configparser==6.0.0
constants==0.6.0
cryptography==41.0.7
decorator==5.1.1
docutils==0.20.1
et-xmlfile==1.1.0
filelock==3.13.1
flex==6.14.1
genson==1.2.2
grpcio==1.59.3
grpcio-tools==1.59.3
h11==0.14.0
idna==3.6
jmespath==1.0.1
jsonpath-ng==1.6.0
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
natsort==8.4.0
numpy==1.26.2
openpyxl==3.1.2
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pbr==6.0.0
platformdirs==3.11.0
ply==3.11
protobuf==4.25.1
pyarrow==14.0.1
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyonemap==0.10.0
pyOpenSSL==23.3.0
pysnowflake==0.1.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
RESTinstance==1.3.0
rfc3987==1.3.8
robotframework==6.1.1
robotframework-assertion-engine==3.0.3
robotframework-browser==18.0.0
robotframework-databaselibrary==1.4.1
robotframework-datadriver==1.9.0
robotframework-excellib==2.0.1
robotframework-jsonlibrary==0.5
robotframework-pabot==2.16.0
robotframework-pythonlibcore==4.3.0
robotframework-requests==0.9.6
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==6.2.0
robotframework-stacktrace==0.4.1
rpds-py==0.13.1
s3transfer==0.8.2
selenium==4.15.2
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
snowflake-connector-python==3.5.0
sortedcontainers==2.4.0
strict-rfc3339==0.7
tomlkit==0.12.3
tornado==6.4
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
validate-email==1.3
wheel==0.42.0
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
